organize components and add index.tsx for simplified API interface
