+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/ChannelControlsList.tsx
@@ -1,7 +1,7 @@
 import { Accordion } from "@mui/material";
 import { AccordionHeader, AccordionDetails } from "@czi-sds/components";
 import cns from "classnames";
-import { ChannelControl, ChannelControlProps } from "./ChannelControl";
+import { ChannelControl, ChannelControlProps } from "./components/ChannelControl/ChannelControl";
 import { ImageLayer, ChannelProps } from "@idetik/core";
 import { useState, useEffect } from "react";
