+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/components/ChannelControl/components/ColorPicker/index.tsx
@@ -1,0 +1,1 @@
+export { ColorPicker } from "./ColorPicker";
