+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/components/ChannelControl/components/ContrastSlider/ContrastSlider.tsx
