+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/components/ChannelControl/components/ContrastSlider/index.tsx
@@ -1,0 +1,1 @@
+export { ContrastSlider } from "./ContrastSlider";
