+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/components/ChannelControl/components/VisibilityToggle/index.tsx
@@ -1,0 +1,1 @@
+export { VisibilityToggle } from "./VisibilityToggle";
