+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/components/ChannelControl/index.tsx
@@ -1,0 +1,2 @@
+export { ChannelControl } from "./ChannelControl";
+export type { ChannelControlProps } from "./ChannelControl";
