+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/ChannelControlsList/index.tsx
@@ -1,0 +1,1 @@
+export { ChannelControlsList } from "./ChannelControlsList";
