+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/Renderer/Renderer.tsx
