+++ packages/react/src/components/viewers/OmeZarrImageViewer/components/Renderer/index.tsx
@@ -1,0 +1,1 @@
+export { Renderer } from "./Renderer";
